refactor: 세션에 필요한 정보만 저장하고, 인터셉터 범위 수정
