+++ src/main/java/roomescape/global/MemberArgumentResolver.java
@@ -1,13 +1,13 @@
 package roomescape.global;
 
 import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.context.request.ServletWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import roomescape.domain.member.Member;
 
 public class MemberArgumentResolver implements HandlerMethodArgumentResolver {
 
@@ -24,7 +24,14 @@ public class MemberArgumentResolver implements HandlerMethodArgumentResolver {
             final WebDataBinderFactory binderFactory
     ) {
         HttpServletRequest request = ((ServletWebRequest) webRequest).getRequest();
-        Member member = (Member) request.getAttribute("LOGIN_MEMBER");
-        return member.getId();
+        HttpSession session = request.getSession(false);
+        if (session == null) {
+            throw new IllegalStateException("로그인이 필요합니다.");
+        }
+        SessionMember member = (SessionMember) session.getAttribute("LOGIN_MEMBER");
+        if (member == null) {
+            throw new IllegalStateException("로그인이 필요합니다.");
+        }
+        return member.id();
     }
 }
