+++ src/main/java/roomescape/global/SessionMember.java
@@ -1,0 +1,6 @@
+package roomescape.global;
+
+import roomescape.domain.member.MemberName;
+
+public record SessionMember(Long id, MemberName name) {
+}
