+++ src/main/java/roomescape/global/interceptor/AdminOnlyInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape.global;
+package roomescape.global.interceptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -6,12 +6,13 @@ import jakarta.servlet.http.HttpSession;
 import org.springframework.web.servlet.HandlerInterceptor;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberRole;
+import roomescape.global.SessionMember;
 import roomescape.repository.MemberRepository;
 
-public class AuthInterceptor implements HandlerInterceptor {
+public class AdminOnlyInterceptor implements HandlerInterceptor {
     private final MemberRepository memberRepository;
 
-    public AuthInterceptor(final MemberRepository memberRepository) {
+    public AdminOnlyInterceptor(final MemberRepository memberRepository) {
         this.memberRepository = memberRepository;
     }
 
@@ -25,18 +26,16 @@ public class AuthInterceptor implements HandlerInterceptor {
         if (session == null) {
             throw new IllegalStateException("로그인이 필요합니다.");
         }
-        Member member = (Member) session.getAttribute("LOGIN_MEMBER");
+        SessionMember member = (SessionMember) session.getAttribute("LOGIN_MEMBER");
         if (member == null) {
             throw new IllegalStateException("로그인이 필요합니다.");
         }
-        Member foundMember = memberRepository.findById(member.getId())
+        Member foundMember = memberRepository.findById(member.id())
                 .orElseThrow(() -> new IllegalStateException("존재하지 않는 멤버입니다."));
-        String uri = request.getRequestURI();
-        if (uri.startsWith("/admin") && member.getRole() != MemberRole.ADMIN) {
+        if (foundMember.getRole() != MemberRole.ADMIN) {
             response.sendError(HttpServletResponse.SC_FORBIDDEN, "관리자 권한이 필요합니다.");
             return false;
         }
-        request.setAttribute("LOGIN_MEMBER", foundMember);
         return true;
     }
 }
